+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Updated `azure` module
 ([#886](https://github.com/census-instrumentation/opencensus-python/pull/886))
+- Updated `azure` module to enable Azure Functions integration
+([#1010](https://github.com/census-instrumentation/opencensus-python/pull/1010))
 - PeriodicMetricTask flush on exit
 ([#943](https://github.com/census-instrumentation/opencensus-python/pull/943))
 - Change blacklist to excludelist